Add hyksort parallel tests
